# docs/swagger.yaml
basePath: /
definitions:
  handler.HealthResponse:
    properties:
      service:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
  handler.LoginRequest:
    properties:
      email:
        example: admin@example.com
        type: string
      password:
        example: admin123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  middleware.ErrorResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  model.EmployeeStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
  service.CreateDepartmentRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
  service.CreateEmployeeRequest:
    properties:
      department_id:
        type: integer
      email:
        maxLength: 150
        type: string
      joined_at:
        type: string
      name:
        maxLength: 150
        minLength: 2
        type: string
      phone:
        maxLength: 20
        type: string
      position:
        maxLength: 100
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.EmployeeStatus'
        enum:
        - active
        - inactive
    required:
    - email
    - name
    - position
    type: object
  service.UpdateDepartmentRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
  service.UpdateEmployeeRequest:
    properties:
      department_id:
        type: integer
      email:
        maxLength: 150
        type: string
      joined_at:
        type: string
      name:
        maxLength: 150
        minLength: 2
        type: string
      phone:
        maxLength: 20
        type: string
      position:
        maxLength: 100
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.EmployeeStatus'
        enum:
        - active
        - inactive
    required:
    - email
    - name
    - position
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: RESTful API for Employee Management System built with Go Fiber, GORM,
    and MySQL.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Employee Management System API
  version: "1.0"
paths:
  /api/audit-logs:
    get:
      description: Fetches system audit activity history (Admin only)
      parameters:
      - description: Entity filter (employee / department)
        in: query
        name: entity
        type: string
      - description: Action filter (create / update / delete)
        in: query
        name: action
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List audit trail logs
      tags:
      - Audit Logs
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password, returning a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: User login
      tags:
      - Auth
  /api/auth/me:
    get:
      description: Returns the profile information of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Auth
  /api/departments:
    get:
      description: Returns list of all departments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all departments
      tags:
      - Departments
    post:
      consumes:
      - application/json
      description: Creates a new department (Admin only)
      parameters:
      - description: Department info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateDepartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a department
      tags:
      - Departments
  /api/departments/{id}:
    delete:
      description: Deletes a department by ID (Admin only)
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete department
      tags:
      - Departments
    get:
      description: Returns single department by ID
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get department by ID
      tags:
      - Departments
    put:
      consumes:
      - application/json
      description: Updates an existing department (Admin only)
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      - description: Department update info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update department
      tags:
      - Departments
  /api/employees:
    get:
      description: Searches, filters, sorts and paginates employees list
      parameters:
      - description: Search query by name, email, position
        in: query
        name: search
        type: string
      - description: Department ID filter
        in: query
        name: department_id
        type: integer
      - description: Status filter (active / inactive)
        in: query
        name: status
        type: string
      - default: created_at
        description: Sort field (name, email, position, status, joined_at, created_at)
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order (asc / desc)
        in: query
        name: sort_order
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List employees
      tags:
      - Employees
    post:
      consumes:
      - application/json
      description: Creates a new employee (Admin only)
      parameters:
      - description: Employee creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an employee
      tags:
      - Employees
  /api/employees/{id}:
    delete:
      description: Deletes an employee by ID (Admin only)
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an employee
      tags:
      - Employees
    get:
      description: Returns employee details by ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employee by ID
      tags:
      - Employees
    put:
      consumes:
      - application/json
      description: Updates an existing employee (Admin only)
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Employee update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an employee
      tags:
      - Employees
  /api/employees/export/csv:
    get:
      description: Generates and streams a CSV file of employee records
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      - description: Department ID filter
        in: query
        name: department_id
        type: integer
      - description: Status filter
        in: query
        name: status
        type: string
      - default: name
        description: Sort by field
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort order
        in: query
        name: sort_order
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV File
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export employees to CSV
      tags:
      - Employees
  /health:
    get:
      description: Returns the health status and UTC timestamp of the server
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.HealthResponse'
      summary: Check server health status
      tags:
      - Health
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"